+++ releasenotes.md
@@ -11,7 +11,7 @@
 * The choice of colors used in the numerous translucent effects in *DOOM Retro* has been improved.
 * Instead of being set to the currently selected item in the corresponding menu, the `episode`, `expansion`, `savegame` and `skilllevel` CVARs are now set to read-only strings of the episode, expansion, save game and skill level for the current game.
 * The player’s path in the automap, enabled using the `am_path` CVAR, will no longer be recorded while the player is in “no clipping mode” using the `IDCLIP`/`IDSPISPOPD` cheat or `noclip` CCMD.
-* The `mapstats` CCMD will now break down the number of things in a map into monsters, pickups and decorations.
+* The number of monsters, pickups and decorations, as well as the number of liquid and damaging sectors, are now displayed in the output of the `mapstats` CCMD.
 
 ---